Warm up the core cache after the final cache clean on update completion
The core cache warmed up during the database update step (UpdateDatabase)
was wiped by the CacheCleaner::cleanFolders() call in UpdateComplete, which
clears var/cache. On PrestaShop 9.0+ the back office is served by Symfony and
relies on that cache, so the freshly updated shop was left with an empty prod
container and the admin entry point fell back to the front office until a
manual cache clear.

Move getCoreUpgrader() to AbstractTask and warm the core cache again after the
final cleanFolders() so the updated shop boots straight into the back office.
